Refactored the system pallet
