+++ runtime/src/configs/mod.rs
@@ -23,3 +23,6 @@ pub use sudo::*;
 
 pub mod grandpa;
 pub use grandpa::*;
+
+pub mod system;
+pub use system::*;
